+++ src/locations.js
@@ -1111,7 +1111,7 @@ function get_location_type_penalty(type, stage, stat, category) {
         parent_location: locations["Nearby cave"],
         first_reward: {
             xp: 520,
-            messages: ["As you descend deeper and deeper, a sudden thought strikes you - what if you instead tried going up?"],
+            messages: ["reward msg go up"],
         },
         repeatable_reward: {
             xp: 260,
@@ -1363,7 +1363,7 @@ There's another gate on the wall in front of you, but you have a strange feeling
             },
             {
                 required_clear_count: 4,
-                messages: ["In a moment of quiet between fights, you can make out a faint sound of rushing water in the distance"],
+                messages: ["reward msg rushing water"],
                 locations: [{location:"Forest den traversal"}]
             }
         ],
@@ -2319,7 +2319,7 @@ There's another gate on the wall in front of you, but you have a strange feeling
         parent_location: locations["Forest road"],
         repeatable_reward: {
             xp: 1000,
-            messages: ["You reach a wall of water and push through"],
+            messages: ["reward msg through the water"],
             locations: [{ location: "Forest lake" }],
             move_to: {location: "Forest lake"}
         },